Track total_playtime
Given a user id and a game id, the playtime for a given game will increase by a random or set amount. When a game is played for a random amount of time no more than two hours of playtime will be added.

A specific game can be played if a game_uuid is provided, otherwise a collection_uuid will need to be provided in order to play a random game from that collection.