Updated Play_Game
Play_game now works straight with tokens and will defualt to splitting the same of a game into a game name and a time to play when presented with the opportunity to do so. This just means that you can't play certain games, like battlefield 2 for a random amount of time.

to play a random game from a collection, give the playtime first and then the collection you want to play